working animation

+++ Project1/Primatives/Vertex.h
@@ -11,7 +11,7 @@ namespace Primative {
 		glm::vec2 tex;
 		glm::vec3 norm;
 		// std::array<Render::Animation::BoneDetails, MAX_BONE_WEIGHTS> boneDetails;
-		glm::ivec4 ids;
+		glm::vec4 ids;
 		glm::vec4 weights;
 		inline Vertex(glm::vec3 p, glm::vec2 t = { 0, 0 }, glm::vec3 n = { 0, 1, 0 }) : pos(p), tex(t), norm(n), ids(-1), weights(0) { };
 	};
